backend/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
